delete firmware file when delete record in firmware.

+++ server/firmwaredelete.php
@@ -10,13 +10,20 @@ if(isset($_GET['version']))
 	if ($version == "") {
 		$error = "版本号为空！";
 	} else {
-		$query = "DELETE from firmware WHERE version = '$version'";
+        $query = " SELECT url from firmware WHERE version = '$version'";
+        $result = queryMysql($query);
+        if ($row = mysqli_fetch_assoc($result)) {
+            unlink("../" . parse_url($row['url'])['path']);
+            $query = "DELETE from firmware WHERE version = '$version'";
 
-		if (queryMysql($query) == true) {
-            $error = "删除成功！";
-		} else {
+            if (queryMysql($query) == true) {
+                $error = "删除成功！";
+            } else {
+                $error = "删除失败！";
+            }
+        }else{
             $error = "删除失败！";
-		}
+        }
 
 		redirect('firmwarelist.php');
 	}
